refactor: 수정된 예약 정보 데이터에 맞게 수정
